Fix analysis failures
